Fix for pattern, transparency, clipper interaction issues
Three issues this fixes.  One is that if a clipper device is installed during the filling
with a transparency tile, we need to make sure we get the pdf14 device to pop the group
and not use the clipper device.

The second issue was that fit_fill was not getting used properly in the tile filling code.

The third issue was that the wrong dimensions were supplied to gx_trans_pattern_fill_rect when
the pattern was clist based.

This should fix Bug 693365.