Add in pdf14 abort compositor action
The fuzzing files have identified that the interpreter was popping the
pdf14 device during a stop.  This caused the pdf14 device to think it
was time to do its put_image command since it was done, causing a lot
of potential segv issues.   This commit adds an abort compositor action
which will clean up the pdf14 device and set it back up as a pass through
device  allowing for a more graceful exit.