Bug 694575: workaround for filenameforall in regression tests
(This is slightly hacky!)

For the regression testing, especially the weekly tests, several of the QL
CET files give variable results depending on the exact content of the current
working directory - because they use filenameforall and vary their output based
whether it returns an error (mainly due to the path ending up too long for the
scratch string).

This commit "patches" the filenameforall operator to catch those cases, and
(hopefully) always return with no files, and no errors, thus giving consistent
results between runs.

Cluster differences in the CET files, as expected.