PDF interpreter - update to BPC change for JPEG2000
Bug #694621

If we find a box, which we shoudl not as this is only supposed to be used
when each colour channel has differing depth, we parse teh BPC from that,
and ignore the BPC in the ihdr box.

If the components are not all equal, we use the first one and raise a
warning.

No differences expected