Knockout non-isolated transparency group support.
Fixes bug 692766

Proper non-isolated knockout group support added.

This commit adds non-isolated knockout group support to gs.  Also, previously, non-isolated groups that
had color space mismatches were being forced to be isolated.  We now color convert the parent buffer
to the new groups color space to ensure that the group is handled in the way it should.

This work revealed an issue in the color handling of planar buffers which was also fixed.

Also, an issue with clipping to the soft mask boundary during a transparency group pop was was
discovered and fixed as well as an overprint simulation issue in transparency.

Several progressions occurred with this fix.

Fixes bug 692766.