Fix Segfaults with --saved-pages.
If the page was not large enough to use Banding mode, the clist_reset
would crash after the print request finished and the page reloaded.
The gx_saved_page_load needs to be after the saved_pages_list is put
back in the device.

Also, the "file_is_new" status was incorrect if pages were saved, but
no printing was done. This was seen with the TIFF devices and psdcmyk.

The psi/imainarg.c has a #define FORCE_SAVED_PAGES_TEST (disabled by
default) to simplify testing on the cluster without modifying command
line parameters.