Bug 694519: improve PCL performance with FAPI
Correctly maintain "design_grid" parameter in the font/matrix pair
structure, for FAPI handled TTFs.

Change a "return" to a "return_error".

Retrieving glyph metrics was sub-optimal as my original implementation used
a gs_text_process() call, which uses the entire "show machinery".

This implementation calls gs_fapi_do_char() directly. This has several benefits:
we only need a "skeleton" text enumerator, and a "skeleton" graphics state,
cutting down the local overhead setup. We also cutout a whole lot of setup
done as part of gs_text_begin() and during gs_text_process() itself. Finally,
we avoid a lot of "stuff" related to showing entire strings, when we're only
ever handling a single glyph at a time.

A final optimisation is in the FAPI/FT code: if we're just retrieving metrics,
we no longer take the action of retrieving the glyph from FT - which is not
a very cheap operation.

This commit causes ~100 cluster tests to have pixel shift differences.