Fix more Segfaults with --saved-pages and psdcmyk and many differences.
Devices that have devn paramaters could close the device so make sure
and open it. Also set banding_type to BandingAlways while in saved-pages
mode and restore it when we end saved-pages mode.

A few files still segfaulted, and there were many differences because
the page was not erased after the saved-pages "begin".

In gx_saved_page_load, for pages that fit in a single band (as with
large -dMaxBitmap values), reset the ymin and ymax to force rasterizing
the next page's clist.