Have pdfwrite ignore WMode=1 when a font has no Metrics2 entry
This arose while investigating a customer bug. Sadly its not actually
anything to do with the customer problem, but it was incorrect. The spec
says that if a font has WMode 1 but has no Metrics2 entry it should be
treated as horizontal (WMode = 0)

Quality Logic file 16-09.ps shows a progression with this change.