Bug 694699: Scale FAPI's font BBox for units.
When FAPI "refined" the FontBBox from the scaler's data, it failed to correctly
take account of the units in which the scaler returned the values.

There was some code which *may* have been an attempt to handle it, but was
wrong in the general case. That's removed in favour of handling it properly.

Several regression files show differences, mainly with pdfwrite/ps2write: many
are pixel differences due to rounding, some are progressions. One was wrong
prior to this commit, and is slightly differently wrong with it.