pdfwrite - don't emit extra blank page with '%d' output file spec
Bug #694576 "Using %d with pdfwrite results in extra, blank page"

Added a flag to the PDF device structure so that we can tell if pdf_close
has been called as the result of outputting a  page. If it has not, and
the page is unmarked, and we are outputting in 'file per page' mode, then
we don't emit the page.

However, this leads to a 0 byte output file, so we also need to delete the
current output file. There seemed to be no obvious way to do this, so I
have added a gx_device_delete_output_file very similar to the existing
gx_device_open_output_file and gx_device_close_output_file routines.

No differences expected.