pdfwrite - check return values from 'convert color'
Bug #694221 "Valgrind issues found by fuzzing in choose_DCT_params (gdevpsdi.c:156)"

In the past convert_color could not fail, with ICC handling, it can, so
we need to check the return code and handle errors gracefully.

This would not in fact have caused problems, but we fix it anyway.