pdfwrite - Improve robustness of type 1 font interpreter
Bug #694214 "Seg faults found by fuzzing in type1_next"

The font is corrupted and one of the glyph instrctions performs a 'return'
without previously executing a subroutine. This caused the font interpreter
to crash.

We now check when executing a return to see if we are not in the main glyph
program, if we are, we return an invalidfont error.

No differences expected