pdfwrite - do not compose font name with CMAp when creating a CIDFOnt
Bug #694721 "Issued with font names in pdfwrite device"

When creating a CIDFont for output to a PDF file, we created the name by
composing the CIDFont name with the CMap originally used with the font.
However, this doesn't make sense, the CIDFont we emit is not composed with
that CMap, its usually composed with an Identity CMap. The CIDFont is
still arranged according to the original CMap, but this is the Ordering
of the font, not the CMap being used with it, and we record that Ordering
(should it be useful for anything) in the CIDSystemInfo.

We now do not add the CMap name to the CIDfont name, which allows Acrobat
to do a 'better' job of copy and paste (apparently it passes the font name
along).

No differences expected