pdfwrite - free Named Images stack before clearing local/global named objects
Bug #694136 "Seg faults found by fuzzing in cos_value_free"

If we stored named images in the initial name space then when we closed
the PDF file we would free and delete the local named objects before we
freed the named images stack. If the named images stack included entries
stored in the local named objects stack then we would attempt to free an
object which was already freed.

So we clear the named images stack first, for added safety we ensure we
clear the NI stack before global named objects are freed too.

No differences expected.