Bug 694734: handle Type 0 font in annotation
The code to generate an appearance from an annotation with an AP, the code
required the FontBBox, but failed to take account that the FontBBox is not
required in a Type 0 font.

We now handle it by grabbing the FontBBox from the descendant font - for
which a FontBBox is mandatory.

No cluster differences.