Make order of --saved-pages-test and --saved-pages= options position independent.
The position of these command line options had been required to be the
last option prior to the first file to run, but this is not how most
other command line options operate. This allows the --saved-pages-test
or --saved-pages=begin option to be placed anywhere in the command line.

The debug option to FORCE_SAVED_PAGES_TEST is now in psi/imain.c
and the OMIT_SAVED_PAGES_TEST is removed.