Allow user to override configure's color index setting
This command line parameter is not listed in the "configure --help" output
as it is really intended for our use, and not generally by end users.

To build with 32 bit colour indices, issue:
./configure --with-large_color_index=0

to build with 64 bit color indices:
./configure --with-large_color_index=1

This works for both Ghostscript and GhostPDL builds.

The only "wrinkle" is that Ghostscript default to:
"--with-large_color_index=1" (so 64 bit indices)

whilst GhostPDL defaults to:
 "--with-large_color_index=0" (so 32 bit indices)

This retains the previous default settings.

No cluster differences.