Fix 694737 - Typeface mask incorrect.
The typeface value is in the first 12 bits and the vendor info in the
next 4 bits of the short value that stores typeface information.  We
were extracting the first 11 bits for the typeface.  This problem is
not associated with a known bug, and in fact, it is quite unlikely to
result in a problem, but it is incorrect.  Thanks to Norbert Janssen
for pointing out the problem.