Ensure that gdev_prn_dev_spec_op forwards to the real device.
When the clist hijacks all the device entries, it attempts to forward
the dev_spec_op calls to the generic printer device. This doesn't
help when we have a device specific one. Only the printer device
knows what the original procs were, so the best we can do is to have
gdev_prn_dev_spec_op forwarding.

This was important as the adjust_bandheight gxdso call was not making
it through to the underlying device.