pdfwrite/ps2write - colour code work
1) Rescale inputs to 0->1 before concretizing if the space is specifically
Lab. Required as the inputs to the ICC profile must be 0->1. gx_remap_color
and its children do this for us, but gx_concretize_color does not.

2) Do not corrupt the device color from the interpreter if we are converting
to a different colour space, use a temporary structure instead.

3) Make sure to use frac2float after concretizing to convert the values.

4) Don't forget to update the saved colour (in the device) after we've
converted colours.

5) Update pdsdf_set_color so that if we are using the new colour code we
get the expected number of output components from the ICC profile attached
to the device, rather than the device 'color_info'. Update all calls to
psdf_set_color so that they pass in the required boolean.

This code is still switched off by default, so no cluster differences will
show up.