pdfwrite - colour code, do not apply range limiting for images
When handling an image in a Lab space, do not reset the range of the space.
We don't need to when preserving an image, and if we use the same space
immediately afterward for linework, the range clamping causes the components
to come out black.

No differences, code is not used yet.