pdfwrite/ps2write colour management, enable the new code
Change the default setting of teh colour management switch so that we
default to using the new code.

If we are creating PDF/A-1, force the old colour management, we can't yet
guarantee that we are getting a version 2 profile.

Add some minimal documentation in ps2pdf.htm, hopefully we will do some
more work in this area before release, and the documentation can be revamped.

This change results in a lot of differences, especially with ps2write. In
the main these are slight shifts in color values caused by using a different
(better) conversion system. However, the new code does a much better job
of handling some kinds of colour spaces, in particular Separation spaces
with ps2write which are now preserved. This leads to a number of progressions
in the test suite.

I've looked at all the differences and I believe they are all either very
minor colour shifts, or progressions.

This is with 'LeaveColorUnchanged', actually converting to different spaces
does result in differences but this can't be cluster tested. I believe these
are all progressions however.