pdfwrite - use /Creator instead of /Author for the XML creator entry
Bug #694735 "PDF/A XMP metadata and DOCINFO synchronization"

We were using the Document Info /Author field for the XML 'creator' which
apparently causes at least one PDF/A validator to take offence. Its not
entirely clear to me from the documentation which one is inteded to be
here, so we'll just use /Creator instead.

No differences expected.