pdfwrite - revert a4de55 and check for parentheses properly instead
Bug #694735 "PDF/A XMP metadata and DOCINFO synchronization"

The commit a4de55 was incorrect, the PDF/A spec clearly states that the
XML creator is based on the Author. However, when an empty string is
present for the Document Info entry, we should not write the string parentheses
into the XML, this is clearly incorrect.


Still no differences expected.