pdfwrite - add transparency Group to substream resources
Bug #691352 "cairo pdf mis-distilled."

This fixes the last remaining problem with this old bug. When we were
creating Form XObjects (because of transparency groups) we were assuming
that the from was a child of the page.

However, groups nested within groups (uncommon, except with Cairo's awful
output) can cause a Form XObject to have children which are also Form
XObjects. In this case we need to have the Form (or other substream) add
the child form to its own Resources rather than inheriting them from the
Page.

This also addresses (most of) Bug #693332 though this still leaves a
dangling ExtGState which needs separate investigation.

No Differences expected