Bug 694718: "reset" MT fonts for each glyph.
Seems some vesions of UFST require the font object to be recreated for each
glyph to work correctly. So when I fixed the optimisation in gxfapi.c to
avoid calling "get_scaled_font()" when the font parameters had not changed, it
exposed the problem with the UFST integration.

To address this, the fapi/ufst code now zeros out the stored scale matrix for
MT fonts, so the code in gxfapi.c always registers the scaling as having
changed. Luckily, recreating the MT font object is not a very expensive
operation.

Also, grab the design units scale from the UFST state structure if the values
haven't been set in the outline structure.

CLUSTER_UNTESTED