pdfwrite - add a Resources dictionary to 'Group' dictionaries
Bug #693322 "Acrobat Preflight tool reports missing XObjects"

Crazy Cairo-constructed files can result in Group dictionaries which use
other resources (in this case an ExtGState) and this wasn't possible in
the existing code. This commit adds a Resources dictionary so that it is
now possible to recreate this.

No differences expected