Bug 690571: deal with memory dev/xdev dimensions being different.
For GV (and ghostview) a memory deivce is used. The problem is now
that most time the width and height of the memory device is not
identical with the xdev device its self. Also extrem large bbox
can lead to problems on systems with less physical memory.

This patch avoids these two problems.

No cluster differences.