Tweak clist line_ptrs data.
The clist has always allocated space for the memory device as if
the line_ptrs were to be included within the allocated space, but
has never actually put the line_ptrs into this space.

Consequently the memory device always allocates space for the
line_ptrs itself. This space is then freed on a close_device call.

This can present problems in NumRenderingThreads mode, when
process_page is used. With the existing code, within clist_render_thread,
the buffer device is setup (and the new line_ptrs block is allocated).
The device is then freed at the end of clist_playback_band
(called from clust_render_rectangle). This nukes the line_ptrs
leaving the get_bits_rectangle call from within the process_fn call
to fail when dereferencing NULL.

The fix here is to amend the clist device structure to include the
offset within the block that the line_ptrs should be allocated at.
This offset can then be used to pass in the address of the line_ptrs
to the setup_buffer_device call. This means the line_ptrs survive
device closing.

This still leaves questions (in my mind at least) as to whether the
handling of the device closures within clist_playback_band is correct
or not.