Fix fpng crash, file has_transparency
In clist_playback_band, it checks to see if there is a compositor as
the target, and needs to free it, but it should not close the device
if its is_open is false, since that would forward the close to the
orig_target, closing the memory device. The pdf14 pop_device had
already done the pdf14_close and set the procs to forward.