Improve collection of color_usage.or bits and track devn colors.
Some of this was in place with cmd_drawing_color_usage, but it was
not used many times, and it just returned 'all'. Also, add the
check for devn colors that are widely used with planar devices
such as psdcmyk.

Also handle the color_usage_array in the setup and teardown of the
BGPrint thread and/or multiple rendering threads.