pdfwrite - free any allocated memory when initialising stream filters
Bug #694758 "Segmentation Fault"

It seems that the arcane image setup sequence used by pdfwrite can result
in us calling stream 'init' functions more than once for the same stream.
Some of the stream filters (Average and Bicubic) allocate some working
memory in the 'init' procedure, but they don't check to see if the
working memory has already been allocated.

This is primarily a cause of memory leaks, though since the memory is
garbage collected it won't (probably) actually leak. However it *does*
seem to confuse the garbage collector when relocating the stream state
structure. We allocate a buffer 9.9Kb but when we come to use it it has
shrunk to a little over 1Kb.

This commit checks the working memory and if its not NULL then we free it
before re-allocating (we might be re-initialising to a different size) which
resolves the problem for me. However because it changes the memory pattern
its possible this is simply masking the problem. Note that this change
assumes that the structure members will be set to NULL when the structure
is allocated (which it always is at the moment).

No differences expected