pdfwrite - when unable to convert a Shading colour space, fallback to image
Bug #688731 "Huge PDF file with ProcessColorModel=/DeviceCMYK and ColorConversionStrategy=/CMYK"

When we can't convert the colour space of a shading dictioanary to satisfy
the ColorConversionStrategy we currently fall back to the rendering method
used by gx_fill_default, which is to degenerate into parallelograms.

This results in hugs and very slow PDF files.

This commit tests the ColorConversionStrategy when we detect an inability
to include a Shading unchanged, and if the strategy is not 'LeaveCOlorUnchanged'
then we use the ps2write approach and render the shading to an image instead.

This produces smaller faster output and the quality can be controlled by
setting resolution with the -r switch allowing a tradeoff between size and
quality.

If we should be unable to embed a shading for other reasons we will still
fall back to the parallelograms method. We may want to avoid this in all
circumstances, but lets wait for an example before we decide.

No differences expected.