pdfwrite - set CIDToGIDMap correctly for PDF/A-2 output
Bug #694769 "Fonts changed from TrueType to CID TrueType and not displaying correctly in output PDF"

We opted not to emit CIDSet for PDF/A-2 since its now documented as
optional and I wasn't able to find a CIDSet which would pass Acrobat's
preflight tool.

However, the code also skipped the correct setup of the CIDToGIDMap and we
can't do that or the output will be incorrect.

This commit makes PDF/A-1 and PDF/A-2 behave the same with respect to the
CIDToGIDMap.

No differences expected, we don't test PDF/A production