Introduce a list of device caches within the clist.
This is only used when NumRenderingThreads > 0, and avoids us recreating
caches every page.

We allocate the cache list with the thread safe allocator. This has the
effect of moving the icc_cache_cl entries out of the thread specific
chunked allocators, and also out of garbage collected memory. We
therefore remove the code to enumerate icc_cache_cl.

Also, this has the effect of fixing a gc enumeration bug where the first
element of the imager state wouldn't have been enumerated.