Fix reference counting for icc_cache_cl in clist.
Currently we only rc_decrement the icc_cache_cl entry in the bg_print
case. We should really be decrementing in all cases, but we get away
with this because the icc_cache_cl entries are allocated in the
rendering threads own chunked allocator which is destroyed after each
run.

In a soon-to-come commit, icc_cache_cl will be moved out of the
thread specific chunked allocator, and these will become real leaks.

The simple fix is to rc_decrement in all cases.