PDF interpreter - more heuristics to work around broken PDF files
Bug #594772 " Regression: missing data reading PDF files starting with d5c42d3a14cee808fd889c14805a69adae8eb5b9"

The PDF file is broken, it has a font with a FontMatrix where one of
the values is 0.-4.

Previously we used .pdftokenerror which would detect
this and convert the value to 0 (observed Acrobat behaviour). However at
some point this was modified so that if we are not in a content stream we
convert into a name object instead.

This doesn't work in the case of a matrix. Unfortunately changing it back
broke *many* PDF files. Almost all of them from SumatraPDF and they are
also badly broken but still....

In essence we want to treat broken numbers as '0' and broken anything else
as a name. So the code now tries to decide if the input looks like a broken
number (contains 0-9, ',', '.' or '-' and nothing else). If it looks like a
number we pass it through .pdftokenerror which will convert it to 0. It is
possible for .pdftokenerror to throw an error, though, so we run it in a
stopped context and if it does, we convert to a name instead. If it doesn't
look like a number, we just convert it to a name.

This causes differences in Bug694429.pdf, but the differences are that more
text is now drawn. The output was wrong before, its wrong now, its merely
wrong in a different way.