Fix SEGVs caused by icc_cache_cl memory space change.
A few commits ago, we moved the icc_cache_cl used for NRT > 0 to the
thread safe memory handler. This meant it was no longer in gc-able
space, so we therefore removed the enumeration of icc_cache_cl from
the clist structure definitions.

Unfortunately I forgot to move the non-NPR related allocations of
icc_cache_cl to the same thread safe handler, meaning that they were
still being allocated in gc-able memory and causing crashes. For
example:

  gs/debugbin/gswin32c.exe -Z@ -o out.pbm -sDEVICE=ppmraw -r300
  -dMaxBitmap=400000000 ../ghostpcl/tests_private/comparefiles/Bug690189c.pdf