pdfwrite - new colour code, emit sRGB as ICCBased
Bug #694780 "incoherent output intent and color space in PDF/A"

sRGB was supposed to be written as ICCBased, but the code has a special
check to see if the ICC profile being used is one of the default ones
(ie its our ICC equivalent of Gray, RGB or CMYK) and if it is, we write
the original colour apace out.

Added a boolean to override this behaviour so that when we emit sRGB we do
in fact write the ICCBased equivalent profile for RGB.

No differences expected.