pcl/FAPI MT font fixes.
Fix a problem with parsing out the FCO paths from the FCO list string (we could
run off the end of the string).

Make the "header" byte swapping call conditional on being a little-endian
platform.

Fix a makefile hiccup that resulted in the FCO files being built into the romfs
twice.

CLUSTER_UNTESTED