Bug 694601: Fix big endian crash in mem_mono_copy_mono.
While the crash is reported on a big endian device, the problem actually
potentially occurs on little endian systems too. As it happens,
x86, amd64, and many ARM configurations will read happily from unaligned
addresses. Some ARMs and many other processors will not, however, so we
apply a fix that will work properly everywhere.

When calling device level functions, we should ensure that the 'raster'
given is always a multiple of align_bitmap_mod bits.

Thanks to John Wehle for the bug report and suggested fix.