pdfwrite - fix dictionary stream hashing
The dictionary stream hashing code didn't initialise the MD5 hash before
using it, which could lead to problems. Also the cos_stream_equal routine
was passing in an MD5 state which would then be reused inside the hashing
code. Although this doesn't seem to cause a problem, and we don't actually
use that state, its safest if we make use a dummy instead.

Problems discovered while working on duplicate shadings.

No differences expected