pdfwrite - spot reuse of shading dictionaries and don't emit multiple times
Bug #689247 "pdfwrite creates redundant shadings"

Previously we always made a new shading and pattern dictionary every time
we encountered a shading. Taking advantage of the hashing mechanism that
was written to improve performance, we can now spot reuse of shading
dictionaries, and emit only a single copy.

The output from the test file now only contains 106 shadings and patterns
and the compressed file size is reduced by 41% as a result.

No differences expected.