Fix bug 689805: Use a clist for pages with transparency that need lots of RAM
Devices that are not gx_device_printer class try to allocate full page bitmaps
for the stack of transparency buffers. This can kill performance or result in
VMerror during transparency stack pushes. Devices that are gx_device_printer
class automatically used clist mode, but other devices did not. Devices that
do not internally support clist mode are recognized by the return from
gxdso_supports_saved_pages device spec_op.

A gx_device_printer device is used to accumulate the page, then when the
PDF14_POP_DEVICE is performed, the page is sent to the original device
as an image.

NB, both the pdf14clist___ device and the pdf14___ device are kept in the
device chain since we have to be able to accomodate unwinding the gs_state
structure that may point to either as various "restore" ops occur.
Unfortunate, but until we come up with a way to fix the chain of compositors,
this is the best we can do.

Move space_params to gx_device struct in order to allow non-printer devices
control over MaxBitmap, BufferSpace, BandBufferSpace, and BandHeight. The x11
device used MaxBitmap somewhat differently, so it now uses the space_params
value, and sets the default in its 'open' function, rather than the device
prototype. The duplicate 'page_uses_transparency' in space_params is removed,
and all devices have this as a parameter.

Also, get rid if the nasty hacks for detecting pattern-clist devices based on
the device name, and export these for the pdf14 device and other clients.

Finally, (unrelated) remove the #if 0 for testing with small memory on
large memory machines, and get rid of the the description of the '.' option
since this hasn't worked for years, and the command line options can readily
be set anyway.