pdfwrite - ignore annotation /AP where the appearance is not a stream
Bug #694809 "Conversion of PDF-file fails with typecheck-error"

The speciment PDF file has an invalid appearance for an annotation:

182 0 obj
<<
/AP << /N << /Yes /null >> >>

The normal appearance should be a dictionary where each key value pair has
a value which is a stream. /null is not a stream.

This commit drops annotations with this problem and emits a warning.

No differences expected.