Add debugging for max heap memory allocation for -Z:
This is a minor hack for debug builds that prints out the 'max' usage
which is often of interest. A better method would be to add a memory
proc to print the stats that would be implemented for all allocators.

If so, a 'verbosity' control might be nice for 'max' (max used) and
'leaked' that would print all elements that the allocator knows about
(for leak detection), or 'current' to print current allocated size.