pdfwrite - add Resources to current substream when starting an SP pdfmark
Bug #689653 "Nested form XObject pdfmarks not working properly"
Bug #689883 "/SP pdfmark inside /BP-/EP pdfmark pair for XObject re-use"

When we use a 'picture' via the SP pdfmark, add the picture resource to
the current substream resources, so that if we are accumulating a substream
(pattern, form, another picture etc) then the Resources for that substream
will contain an entry for the picture. Previously they did not and relied
on inheriting from the page dictionary, which is long since deprecated.

Thanks to Alexander Grahn for diagnosing the problem and supplying the
patch.

No differences expected.