pdfwrite - fix compiler comparison oddness
Bug #694838 "Artifex Ghostscript 9.10: JBIG2Encode requires CompabilityLevel >= 1.4"

The code compares a float variable with a non-integer constant. Bizzarely
the comparison fails unless the contant is declared as type float as well
(using the x.yf syntax), at least on the Microsoft compiler. Presumably
without this they are evaluated as double.

This commit simply alters the constants to be declared specifically as
floats.

No differences expected.