Bug 694836: remove erroring CIDFont subsitute record
When subsituting a CIDFont we dynamically update the CIDFont resource's .map
dictionary (to ensure the substitute CIDFont is loaded with the correct
"attributes").

If findresource fails with the updated .map dictionary, remove the entry we
just added, to avoid subsequent attempted loads of that CIDFont finding the
erroring .map entry.

No cluster differences.