Bug 694832: refine when to elide the TTF notdef glyph
Only force the Truetype notdef glyph to be non-marking when the font in
question uses a TTF cmap table other than the symbolic (3.0) cmap table.

Several cluster files show differences - all except one are progressions
with reference to Acrobat, and the exception was "wrong" before and is
slightly differently wrong now (an entirely different underlying issue).