Bug 694843: reset FAPI font data on "refine" font
For example, a font dict may be copied, so the "FAPI" entry is copied with
the dict, but the new internal font object does not have the FAPI server in
it. We catch that in the "refine font" call - previously here we didn't reset
the font data, just "re-bound" the FAPI server. It turns out with the
language switch build, we need to set the FAPI font data, too.

No cluster differences.