Bug 694842: Fix threshold table generation.
When we generate threshold tables for dithering, we should expect them
to be 'evenly distributed'. This is not the case currently. In the
example given in this bug, we generate an 8x4 dither table. This
contains 32 pixels, all of which can be on or off, hence we have 33
levels, ranging from 0 on to 32 on.

We would further expect our input range to be evenly divided into these
levels - i.e. roughly 256/33 in each band.

This is not the case. The threshold table generated is:

threshold array row   0=  88 120 152 184 176 144 112  80
threshold array row   1= 192 224 208 168  72  40  56  96
threshold array row   2= 232 255 240 136  32   8  24 128
threshold array row   3= 216 248 200 104  48  16  64 160

The operation of our thresholding code is to set a pixel if
contone(x,y) < threshold(x,y), hence our inputs divide into:
0-7,8-15,...240-254,255

Not only does this not have the required 256/33 per region, the top
and bottom regions are far from being even (8 vs 1 entries).

Ideally we should fix this code so generate the correct distribution -
an enhancement bug will be opened for this. For now we will satisfy
ourselves by offsetting the thresholds slightly so that the top and
bottom ranges are at least even. The new code gives:

threshold array row   0=  84 116 148 180 172 140 108  76
threshold array row   1= 188 220 204 164  68  36  52  92
threshold array row   2= 228 252 236 132  28   4  20 124
threshold array row   3= 212 244 196 100  44  12  60 156