Bug 694844: Fix thin stroking of paths with non-idempotent colors.
For non-idempotent colors when stroking we have to construct an entire
path and then fill it, rather than stroking each section as we go.

Unfortunately in the code that implements this, we missed a case - thin
sections can be stroked section by section even in the non-idempotent
case.

We fix that here.

This appears to solve Bug 694844, which was hitting this case.