ps2write - Don't emit a page size change if the last request failed (and was the same size)
Bug #694852 We check the current media size at the start of every page, and
only emit a new request if the page is a different size to the current size.

However, if the request fails (media not available) we keep on requesting
the media via setpagedevice. This causes duplexing to be disabled, as each
setpagedevice resets the device and ejects the previous page.

This commit still checks the current page size, however, if it is not the
size needed for the page, we then check to see if 'MediaRequested;' is
known. If it is not we set the media, and set MediaRequested to the new
media size. If MediaRequested is known, we check to see if the new media
size is the same as the last one we requested. If it is there's no point
in doing a new request as it will fail again,so we drop the request.

If the new requested media is not the same as the last requested media,
then we do emit a new request, which may or may not be satisfied of course.

This should permit duplexing to work and be slightly faster in the case
where requested media is not available.