pdfwrite - don't allocate multiple object IDs for shadings
Bug #694853 "PDF/A generation does not fix all errors"

When the code to detect and elide duplicate shadings was added we used the
pdf_substitute_resource function to detect duplicates and drop the new one.

However this function *always* assigned a new object ID if it wasn't a
duplicate, which leaves the original ID dangling (if the object already had
one). This commit simply avoids creating a new ID if the object is not a
duplicate, and already has an ID.

I'm a little nervous about this one, but a cluster push seems to be OK. No
differences expected.