Added "pwgraster" output device
The "pwgraster" output device rasterizes the input into the PWG Raster
format, an industry-standarized raster format for printers, especially
the upcoming IPP Everywhere network printers which are supposed to
work with any hardware-model-specific software (drivers). As PWG
Raster is derived from CUPS Raster it is generated by the "cups"
output device, but calling it with the "pwgraster" name makes it
generating PWG Raster instead of CUPS Raster.