Fix bug 69365: Write pattern colors to all bands consistently for images.
Thanks to Michael Vrhel for tracking this down. If a transparent pattern
color is written to all bands after having been written to a band for
an image, the transparency fill_trans_buffer will be NULL because the
group for the image was pushed at 'begin_typed_image' time. We now write
the color to all_bands based on the image extent, consistently.