Enable the pdf14 optimization. Previous commit only collected data.
Somehow, the previous commit that collected the data didn't actually
skip pdf14 transparency on any bands. Also testing showed that fiils
with patterns that needed transparency were not being detected.

We need to save the page mode pdf14_needed state based on the blend_mode
and opacity to restore to the correct setting after smask or transpacrency
group pushes.

Lastly, we punt if the device is not contone, since the colors written
for bands will be contone which are not compatible with the target.

Correct the pdf14_opmode_names to add in the missing PDF14_ABORT_DEVICE.