pdfwrite - add a warning if -dUseCIEColor is set
With the current color management its not really sensible to set UseCIEColor
(which was always a horrible hack anyway) in order to do color conversion
or management.

So if we see this being done, show the user a warning. We don't want to
raise an error as someone will undoubtedly complain that they have a good
reason for using UseCIEColor with pdfwrite.....

No differences expected.