DSC parser - reset scanning at end of job
Currently the DSC parser never leaves 'scan_trailer', which is fine for
one-job-at-a-time usage, but not good for job server or multiple inputs.

This commit modifies the scan-trailer condition so that if a new DSC header
is encountered, *after* we've found a trailer, we reset some aspects of the
DSC parsing.

Note that this doesn't completely reset the parser as that caused memory
problems.

Bug #692208

No differences expected