Don't attempt to copy a missing glyph when copying fonts
Bug #694236 for me does not (initially at least) cause a Valgrind warning,
but does cause a seg fault.

The problem is that we have a (presumably broken) TrueType font where we
try to enumerate all the glyphs in the font. The enumeration code returns
a GS_NO_GLYPH, but the font copying code doesn't catch that, and so tries
to copy that glyph. This causes a crash in the name lookup.

The commit here simply adds an additional check that the glyph is not
'GS_NO_GLYPH', if it is we treat it as an error.

No differences expected.