pdfwrite - limit scan line comparison to actual width
Bug #694930 "Valgrind issues found by fuzzing in compute_subimage (gdevpdfd.c:707)"

The local image converter, and specifically mask to clip conversion,
compare consecutive scan lines of an image to see if they are identical.

However, they were using the width of the allocated raster, which can include
alignment padding, rather then the actual width of the image data. This
meant we were comparing uninitialised data. In this case its actually safe
as we would simply create a more complex clip than required, but its
definitely sub-optimal.

This commit changes the comparison to use the number of bits in the image,
rounded up to the nearest byte.

No differences expected.